Module ssmp_client::parser::printer

Constants

PROMPT_COLOR

USERNAME_COLORS
A constant used in print_message() to print out usernames as different colors.

Functions

color_print
Inner function that prints out colored text. If the write fails, the text is instead printed using print!()

print_error
Prints a standardized error message. Takes in the message to print and prints it with a special prefix.

print_help_menu_item
Prints a single line of the 'help' command's menu. Takes in the name of a command and its description.

print_intro
Prints an 'intro' banner when the client is launched

print_message
Prints a single line (representing a send/received message) formatted like so: <sender_name>: <message>

print_prompt
Prints a prompt for the user.

print_usage
Prints a standardized usage messages. Takes in the message to print and prints it with a special prefix. It also takes in an action, which is a short description of what the command will do. action can be left as "" if none is to be printed

print_username
Takes in a username and prints it a certain color (decided using modular arithmetic).