Module ssmp_client::parser::stubs

Functions

bootstrap

connect
Talks with the server and attempts to connect a user. Returns the IP address, port, and public key of the user on success.

friends_add
Takes in a Friend struct and attempts to add it to the local friend list

friends_remove
Takes in a username (&str) and attempts to remove them from the friend list.

friends_show
Simple stub that prints the contents of the client's friend list.

friends_wipe

get_online_users

history
Takes in a FriendList reference and a username. The friends list is searched for a friend with the matching name. If one is found, a vector of Message structs are returned. Otherwise, an error message is returned.

is_auth

login
Login to the specified user account

logout

msg_query

msg_send

signup