Struct ssmp_client::storage::friends::FriendList

pub struct FriendList {
    size: u32,
    db_table_names: Vec<String>,
    db: FriendDatabase,
}

Friend List

The FriendList struct serves as an interface to the FriendDatabase defined in friend_database.rs. This struct supports common operations that can be called without directly interacting with the database.

Every FriendList struct uses the same database. (Multiple FriendList struct instances won't write friends to different database.)

Fields
size: u32db_table_names: Vec<String>db: FriendDatabase
Implementations

impl FriendList

pub fn new() -> FriendList

Constructor: creates and returns a FriendList pointed at the default database location.

pub fn new_at_location(location: &str) -> FriendList

Constructor: creates and returns a FriendList pointed at a database at the location specified.

fn setup_database(&self) -> bool

Sets up the initial table(s) for FriendSchema. Returns false if an error occurred while setting things up.

pub fn add_friend(&self, friend: &Friend) -> bool

Attempts to add the given friend to the friend list. Returns true on success and false on internal error.

pub fn find_all(&self) -> Option<Vec<Friend>>

Searches for ALL friends in the friend list. Returns a vector of Friend structs on success, and None on failure.

pub fn find_by_username(&self, username: &str) -> Option<Friend>

Searches the friend list for a friend with a matching username. If one is found, a Friend struct is returned. Otherwise, None is returned.

pub fn find_by_public_key(&self, pubkey: &PublicKey) -> Option<Friend>

Searches the friend list for a friend with a matching public key. If one is found, a Friend struct is returned. Otherwise, None is returned.

pub fn find_by_ip_address(&self, address: &str) -> Option<Vec<Friend>>

Searches the friend list for friends with a matching IP address. Since multiple friends could have the same IP address, this function returns a Vec<Friend> on success, rather than just a Friend. On failure, it returns None.

pub fn update_friend_public_key(
&self,
username: &str,
new_key: &PublicKey
) -> bool

Takes in a username and PublicKey and attempts to update the entry in the friend list with the matching username. Returns true on success and false on failure.

pub fn update_friend_ip_address(&self, username: &str, new_addr: &str) -> bool

Takes in a username and IP address (both as strings) and attempts to update the entry in the friend list with the matching username. Returns true on success and false on failure.

pub fn remove_friend(&self, username: &str) -> bool

Attempts to remove the friend with the given username from the friend list. Returns true on a successful removal and false otherwise.

pub fn wipe(&self) -> bool

Attempts to clear the entire friend list. (The SQLite database used to store the friends is completely deleted.) Returns true on a successful wipe and false on a failure.