Struct ssmp_client::storage::message_database::MessageSchema

pub struct MessageSchema {
    pub message_id: (i32, SQLColumnInfo),
    pub conversation_id: (i32, SQLColumnInfo),
    pub message: (String, SQLColumnInfo),
    pub sender: (String, SQLColumnInfo),
    pub timestamp: (i64, SQLColumnInfo),
    pub is_read: (bool, SQLColumnInfo),
}

The Message Schema

A schema used to organize the data kept in the SQLite message database.

Fields
message_id: (i32, SQLColumnInfo)conversation_id: (i32, SQLColumnInfo)message: (String, SQLColumnInfo)sender: (String, SQLColumnInfo)timestamp: (i64, SQLColumnInfo)is_read: (bool, SQLColumnInfo)
Implementations

impl MessageSchema

pub fn new(msg: &str, sender: &str) -> MessageSchema

Constructor: takes in a message (String) and a sender (String) and creates a message schema with the other fields filled in with defaults. The schema's timestamp is set to the current time.

pub fn to_sql_params(&self) -> Vec<&dyn ToSql>

Creates a vector of special ToSql values for SQL commands to parse. The vector contains every struct field in the order they appear.

pub fn to_string(&self) -> String

Creates and returns a string representation of a MessageSchema.

pub fn get_timestamp_as_datetime(&self) -> NaiveDateTime

Takes the schema's timestamp (i64) and returns an appropriate NaiveDateTime object.

pub fn get_timestamp_as_string(&self) -> String

Takes the schema's timestamp (i64) and returns an appropraite NiaveDateTime object

pub fn get_sql_column_info(&self) -> Vec<SQLColumnInfo>

Returns the SQLColumnInfo of each struct field, all combined in a vec.

fn get_current_datetime_i64() -> i64

Gets and returns the current system time as a 64-bit unsigned integer.

fn i64_to_datetime(seconds: i64) -> NaiveDateTime

Takes in a u64 (representing seconds since Jan. 1st, 1970) and returns a NaiveDateTime object with the proper time set.

fn datetime_to_string(timestamp: NaiveDateTime) -> String

Takes the given NaiveDateTime timestamp and converts it to a readable string. The string is returned.